create index.ts in functions folder, articlelist.scss, edit formatDate
