--- exercise-reactjs/src/app/core/functions/replaceText.ts
@@ -1,4 +1,0 @@
-export const replaceText = (text: string, value: string) => {
-  return text.replace('{value}', value);
-}
-
