+++ exercise-reactjs/src/stylesheet/pages/_all.scss
@@ -1,1 +1,2 @@
-@import './article';
+@import './articleList';
+@import './articleDetail';
